Moving hooks into right folder
